+++ kompendium-core/src/test/kotlin/org/leafygreens/kompendium/KontentTest.kt
@@ -7,9 +7,11 @@ import kotlin.test.assertFailsWith
 import kotlin.test.assertNotNull
 import kotlin.test.assertTrue
 import org.leafygreens.kompendium.Kontent.generateKontent
+import org.leafygreens.kompendium.models.oas.DictionarySchema
 import org.leafygreens.kompendium.models.oas.FormatSchema
 import org.leafygreens.kompendium.models.oas.ObjectSchema
 import org.leafygreens.kompendium.models.oas.ReferencedSchema
+import org.leafygreens.kompendium.util.ComplexRequest
 import org.leafygreens.kompendium.util.TestInvalidMap
 import org.leafygreens.kompendium.util.TestNestedModel
 import org.leafygreens.kompendium.util.TestSimpleModel
@@ -19,15 +21,13 @@ import org.leafygreens.kompendium.util.TestSimpleWithList
 import org.leafygreens.kompendium.util.TestSimpleWithMap
 import org.leafygreens.kompendium.util.TestWithUUID
 
+@ExperimentalStdlibApi
 internal class KontentTest {
 
   @Test
   fun `Unit returns empty map on generate`() {
-    // when
-    val clazz = Unit::class
-
     // do
-    val result = generateKontent(clazz)
+    val result = generateKontent<Unit>()
 
     // expect
     assertTrue { result.isEmpty() }
@@ -35,11 +35,8 @@ internal class KontentTest {
 
   @Test
   fun `Primitive types return a single map result`() {
-    // when
-    val clazz = Long::class
-
     // do
-    val result = generateKontent(clazz)
+    val result = generateKontent<Long>()
 
     // expect
     assertEquals(1, result.count(), "Should have a single result")
@@ -47,41 +44,25 @@ internal class KontentTest {
   }
 
   @Test
-  fun `Throws an error when top level generics are detected`() {
-    // when
-    val womp = mapOf("asdf" to "fdsa", "2cool" to "4school")
-    val clazz = womp::class
-
-    // expect
-    assertFailsWith<IllegalStateException> { generateKontent(clazz) }
-  }
-
-  @Test
   fun `Objects reference their base types in the cache`() {
-    // when
-    val clazz = TestSimpleModel::class
-
     // do
-    val result = generateKontent(clazz)
+    val result = generateKontent<TestSimpleModel>()
 
     // expect
     assertNotNull(result)
     assertEquals(3, result.count())
-    assertTrue { result.containsKey(clazz.simpleName) }
+    assertTrue { result.containsKey(TestSimpleModel::class.simpleName) }
   }
 
   @Test
   fun `generation works for nested object types`() {
-    // when
-    val clazz = TestNestedModel::class
-
     // do
-    val result = generateKontent(clazz)
+    val result = generateKontent<TestNestedModel>()
 
     // expect
     assertNotNull(result)
     assertEquals(4, result.count())
-    assertTrue { result.containsKey(clazz.simpleName) }
+    assertTrue { result.containsKey(TestNestedModel::class.simpleName) }
     assertTrue { result.containsKey(TestSimpleModel::class.simpleName) }
   }
 
@@ -89,11 +70,10 @@ internal class KontentTest {
   fun `generation does not repeat for cached items`() {
     // when
     val clazz = TestNestedModel::class
-    val initialCache = generateKontent(clazz)
-    val claxx = TestSimpleModel::class
+    val initialCache = generateKontent<TestNestedModel>()
 
     // do
-    val result = generateKontent(claxx, initialCache)
+    val result = generateKontent<TestSimpleModel>(initialCache)
 
     // expect TODO Spy to check invocation count?
     assertNotNull(result)
@@ -104,85 +84,93 @@ internal class KontentTest {
 
   @Test
   fun `generation allows for enum fields`() {
-    // when
-    val clazz = TestSimpleWithEnums::class
-
     // do
-    val result = generateKontent(clazz)
+    val result = generateKontent<TestSimpleWithEnums>()
 
     // expect
     assertNotNull(result)
     assertEquals(3, result.count())
-    assertTrue { result.containsKey(clazz.simpleName) }
+    assertTrue { result.containsKey(TestSimpleWithEnums::class.simpleName) }
   }
 
   @Test
   fun `generation allows for map fields`() {
-    // when
-    val clazz = TestSimpleWithMap::class
-
     // do
-    val result = generateKontent(clazz)
+    val result = generateKontent<TestSimpleWithMap>()
 
     // expect
     assertNotNull(result)
     assertEquals(5, result.count())
     assertTrue { result.containsKey("Map-String-TestSimpleModel") }
-    assertTrue { result.containsKey(clazz.simpleName) }
+    assertTrue { result.containsKey(TestSimpleWithMap::class.simpleName) }
 
-    val os = result[clazz.simpleName] as ObjectSchema
+    val os = result[TestSimpleWithMap::class.simpleName] as ObjectSchema
     val expectedRef = ReferencedSchema("#/components/schemas/Map-String-TestSimpleModel")
     assertEquals(expectedRef, os.properties["b"])
   }
 
   @Test
   fun `map fields that are not string result in error`() {
-    // when
-    val clazz = TestInvalidMap::class
-
     // expect
-    assertFailsWith<IllegalStateException> { generateKontent(clazz) }
+    assertFailsWith<IllegalStateException> { generateKontent<TestInvalidMap>() }
   }
 
   @Test
   fun `generation allows for collection fields`() {
-    // when
-    val clazz = TestSimpleWithList::class
-
     // do
-    val result = generateKontent(clazz)
+    val result = generateKontent<TestSimpleWithList>()
 
     // expect
     assertNotNull(result)
     assertEquals(6, result.count())
     assertTrue { result.containsKey("List-TestSimpleModel") }
-    assertTrue { result.containsKey(clazz.simpleName) }
+    assertTrue { result.containsKey(TestSimpleWithList::class.simpleName) }
   }
 
   @Test
-  fun `generics as enums throws an exception`() {
-    // when
-    val clazz = TestSimpleWithEnumList::class
+  fun `Can parse enum list as a field`() {
+    // do
+    val result = generateKontent<TestSimpleWithEnumList>()
 
     // expect
-    assertFailsWith<java.lang.IllegalStateException> { generateKontent(clazz) }
+    assertNotNull(result)
   }
 
   @Test
   fun `UUID schema support`() {
-    // when
-    val clazz = TestWithUUID::class
-
     // do
-    val result = generateKontent(clazz)
+    val result = generateKontent<TestWithUUID>()
 
     // expect
     assertNotNull(result)
     assertEquals(2, result.count())
     assertTrue { result.containsKey(UUID::class.simpleName) }
-    assertTrue { result.containsKey(clazz.simpleName) }
+    assertTrue { result.containsKey(TestWithUUID::class.simpleName) }
     val expectedSchema = result[UUID::class.simpleName] as FormatSchema
     assertEquals(FormatSchema("uuid", "string"), expectedSchema)
+  }
+
+  @Test
+  fun `Generate top level list response`() {
+    // do
+    val result = generateKontent<List<TestSimpleModel>>()
+
+    // expect
+    assertNotNull(result)
+  }
+
+  @Test
+  fun `Can handle a complex type`() {
+    // do
+    val result = generateKontent<ComplexRequest>()
+
+    // expect
+    assertNotNull(result)
+    assertEquals(7, result.count())
+    assertTrue { result.containsKey("Map-String-CrazyItem") }
+    val ds = result["Map-String-CrazyItem"] as DictionarySchema
+    val rs = ds.additionalProperties as ReferencedSchema
+    assertEquals(ReferencedSchema("#/components/schemas/CrazyItem"), rs)
   }
 
 }